Remove our manual submodule stuff
Per the below links the git module itself
supports getting git submodules recursvily.

* https://github.com/ansible/ansible/pull/3292/files
* https://github.com/ansible/ansible/issues/3291
* https://groups.google.com/forum/#!topic/ansible-project/L2QJadpZNE8

Thus we can remove this task and everything should
still function as expected

NOTE: Suggested action
- Keep this in a branch and make prod2 pull from said branch
- Update mediawiki-core and see if it is upd on prod2
- Update an extension we have branched and see if prod2 upds
- If the above work then we can deploy this change

Further Note:
Currently the state of mediawiki-core causes conflicts on prod2
and the prod2 copy of mediawiki-core is always in a touched and
dirty state. We should fix this..