Add mw vars files and ext + skins tasks
This change leaves a few TODOs in the vars
file for things that are still missing.
These miight need to be split into correct repos.

Also on the TODO is making the extensions track
branches where appropriate.?

This is still an idea and a work in progress!

Before this is tested we also need to create a
branch of the current version of MW that doesnt
have any skins or extensions in it or submodules
pointing to them!