Disabling ParserFunctions briefly
This is a brief global disabling of ParserFunctions. After analysis of the issues noted in the past few days, it seems to have recurring magic word issues that cannot be fixed by rebuilding the localisation cache. I am going to attempt a brief disabling of this extension, after which it will be reenabled after seeing if this fixes the problems. If so, I plan to change the branch of the ParserFunctions extension to the master branch for now to see if that does not address the problems.